workspace/rag-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.9
openai==1.45.0
chromadb==0.5.5
langchain==0.3.7
langchain-openai==0.2.6
langchain-community==0.3.5
langchain-chroma==0.1.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
httpx==0.27.2
